Fix hook restart bug
Previously, when hooks would fail, the `HookSupervisor` would
restart the hook, but the `HookServer` would not have knowledge
of this and would therefore have no `hooks` entries even though
one in fact existed. This fix will make sure that hooks get
restarted, and that the `HookServer` is also aware of them,
as it will restart them itself.